refactor(MainLayout): DrawerBar => Main Layout
