--- packages/client/src/components/drawerBar/index.tsx
@@ -1,1 +1,0 @@
-export { DrawerBar } from './drawerBar'
