+++ packages/client/src/layouts/Main/icons/AvatarIcon.tsx
