+++ packages/client/src/layouts/Main/icons/Forum.tsx
