+++ packages/client/src/layouts/Main/icons/LeftArrow.tsx
