+++ packages/client/src/layouts/Main/icons/Winner.tsx
