+++ packages/client/src/layouts/index.ts
@@ -1,0 +1,1 @@
+export { MainLayout as Main } from './Main'
